[fix] float type to int

+++ work-list/biggest/big_value.h
@@ -1,6 +1,6 @@
 #ifndef BIG_VALUE_H
 #define BIG_VALUE_H
 
-int get_biggets_value(float x, float y, float z);
+int get_biggets_value(int x, int y, int z);
 
 #endif
