+++ work-list/biggest/get_biggest_value.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 
-int get_biggets_value(float x, float y, float z) {
+int get_biggets_value(int x, int y, int z) {
   return x >= y && x >= z
   ? x
   : y >= x && y >= z
